+++ src/routes/(v2)/v2/discriminated-union/schema.ts
@@ -1,0 +1,11 @@
+import { z } from 'zod';
+
+export const schema = z.discriminatedUnion('type', [
+	z.object({
+		type: z.literal('empty')
+	}),
+	z.object({
+		type: z.literal('extra'),
+		roleId: z.string()
+	})
+]);